Fix: Move model about video to video package
