+++ core/model/src/main/kotlin/jp/shirataki707/yamato/core/model/data/video/DetailVideoResources.kt
@@ -1,7 +1,7 @@
-package jp.shirataki707.yamato.core.model.data
+package jp.shirataki707.yamato.core.model.data.video
 
 import android.os.Parcelable
-import jp.shirataki707.yamato.core.model.data.VideoResources.VideoCarouselBlock.VideoSummary
+import jp.shirataki707.yamato.core.model.data.video.VideoResources.VideoCarouselBlock.VideoSummary
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
