+++ feature/home/detail/src/main/kotlin/jp/shirataki707/yamato/feature/home/detail/ui/DetailPageViewModel.kt
@@ -9,13 +9,13 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import jp.shirataki707.yamato.core.common.network.Dispatcher
 import jp.shirataki707.yamato.core.common.network.YamatoDispatchers
 import jp.shirataki707.yamato.core.data.repository.VideoResourceRepository
-import jp.shirataki707.yamato.core.model.data.DetailPageConfig
-import jp.shirataki707.yamato.core.model.data.DetailVideoResources
-import jp.shirataki707.yamato.core.model.data.VideoResources.VideoCarouselBlockType.Channel
-import jp.shirataki707.yamato.core.model.data.VideoResources.VideoCarouselBlockType.Latest
-import jp.shirataki707.yamato.core.model.data.VideoResources.VideoCarouselBlockType.Mountain
-import jp.shirataki707.yamato.core.model.data.VideoResources.VideoCarouselBlockType.Popular
-import jp.shirataki707.yamato.core.model.data.VideoResources.VideoCarouselBlockType.Recommended
+import jp.shirataki707.yamato.core.model.data.video.DetailPageConfig
+import jp.shirataki707.yamato.core.model.data.video.DetailVideoResources
+import jp.shirataki707.yamato.core.model.data.video.VideoResources.VideoCarouselBlockType.Channel
+import jp.shirataki707.yamato.core.model.data.video.VideoResources.VideoCarouselBlockType.Latest
+import jp.shirataki707.yamato.core.model.data.video.VideoResources.VideoCarouselBlockType.Mountain
+import jp.shirataki707.yamato.core.model.data.video.VideoResources.VideoCarouselBlockType.Popular
+import jp.shirataki707.yamato.core.model.data.video.VideoResources.VideoCarouselBlockType.Recommended
 import jp.shirataki707.yamato.core.ui.common.ParcelableResult
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineExceptionHandler
