+++ feature/home/detail/src/main/kotlin/jp/shirataki707/yamato/feature/home/detail/ui/section/DetailInitialSectionState.kt
@@ -1,7 +1,7 @@
 package jp.shirataki707.yamato.feature.home.detail.ui.section
 
 import androidx.compose.runtime.Composable
-import jp.shirataki707.yamato.core.model.data.DetailVideoResources
+import jp.shirataki707.yamato.core.model.data.video.DetailVideoResources
 import jp.shirataki707.yamato.core.ui.common.ParcelableResult
 
 internal data class DetailInitialSectionState(
