+++ feature/home/detail/src/main/kotlin/jp/shirataki707/yamato/feature/home/detail/ui/section/DetailLoadedSectionState.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
 import jp.shirataki707.yamato.core.common.utils.VideoUtils.openYoutubeVideo
-import jp.shirataki707.yamato.core.model.data.DetailVideoResources
+import jp.shirataki707.yamato.core.model.data.video.DetailVideoResources
 
 internal data class DetailLoadedSectionState(
     val detailVideoResources: DetailVideoResources,
