+++ feature/home/main/src/main/kotlin/jp/shirataki707/yamato/feature/home/main/navigation/HomeNavigation.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
-import jp.shirataki707.yamato.core.model.data.DetailPageConfig
+import jp.shirataki707.yamato.core.model.data.video.DetailPageConfig
 import jp.shirataki707.yamato.feature.home.main.ui.HomePageHost
 import kotlinx.serialization.Serializable
