+++ feature/home/main/src/main/kotlin/jp/shirataki707/yamato/feature/home/main/ui/section/HomeInitialSectionState.kt
@@ -2,7 +2,7 @@ package jp.shirataki707.yamato.feature.home.main.ui.section
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import jp.shirataki707.yamato.core.model.data.VideoResources
+import jp.shirataki707.yamato.core.model.data.video.VideoResources
 import jp.shirataki707.yamato.core.ui.common.ParcelableResult
 
 internal data class HomeInitialSectionState(
